requirements.txt
aiogram==3.12.0
aiogram-dialog==2.2.0
redis==5.0.1
environs==11.0.0
sqlalchemy==2.0.43
alembic==1.13.1
psycopg[binary,pool]==3.2.9
# asyncpg==0.29.0  # Временно отключаем из-за проблем с Python 3.13
marshmallow==4.0.1
gspread==6.1.4
google-auth==2.35.0